Bugfix rainbow_on_off mode to use different start colours.
Previously, the lack of an infinite loop meant the start colour
was reset each time the rainbow_on_off function ran, so each side
always had the same sequence of colours.

This patch adds an infinite loop, so that the colour state is
carried over properly.

Interestingly, this reduces the program size by 36 bytes.